requirements.txt
pandas
pyyaml
matplotlib
# openai  # optional, only if you use the OpenAI provider